+++ ClientApp/src/components/campaigns/create-campaign.scss
@@ -4,7 +4,3 @@
   }
 }
 
-button {
-  margin-bottom: 0.5em;
-  background-color: gold;
-}
