+++ ClientApp/src/components/layout/nav-menu/nav-menu.scss
@@ -1,0 +1,58 @@
+﻿#app-bar {
+  background-color: #333;
+  
+  #nav-bar-title {
+    flex-grow: 1;
+  }
+  
+  a {
+    float: left;
+    color: #f2f2f2;
+    text-align: center;
+    padding: 14px 16px;
+    text-decoration: none;
+    transition: background-color, color, 0.4s;
+
+    &:hover {
+      background-color: #dddddd75;
+      cursor: pointer;
+      transition: background-color, color, 0.2s;
+    }
+
+    &.active {
+      background-color: rgb(238, 145, 6);
+      color: #333;
+    }
+  }
+  
+  #menu-button {
+    color: whitesmoke;
+  }
+}
+
+#navbar-account-menu {
+  text-align: center;
+  
+  li {
+    display: flex;
+    justify-content: center;
+  }
+  
+  a {
+    text-decoration: none;
+    color: #333333;
+    p {  
+      font-weight: 200;
+    }
+    
+    &:hover {
+      p {
+        font-weight: normal;
+      }
+    }
+  }
+
+  #navbar-account-menu-header {
+    padding: 10px;
+  }
+}
