+++ ClientApp/src/components/layout/nav-menu/nav-menu.test.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import * as ReactRedux from 'react-redux';
 import { configureStore } from '@reduxjs/toolkit';
 import { StaticRouter } from 'react-router';
-import * as authStore from "../auth/auth-store.slice";
+import * as authStore from "../../auth/auth-store.slice";
 
 import { NavMenu } from "./nav-menu";
 import { Profile } from "oidc-client";
