+++ ClientApp/src/components/layout/sidebar/sidebar-category.tsx
