+++ ClientApp/src/components/layout/sidebar/sidebar.scss
